lab5-timeseries/requirements.txt
pandas
numpy
scikit-learn
matplotlib
requests
kagglehub
psycopg2-binary
